+++ mirar/processors/skyportal/__init__.py
@@ -2,5 +2,6 @@
 Central module for the skyportal integration
 """
 from mirar.processors.skyportal.client import SkyportalClient
-from mirar.processors.skyportal.send_to_skyportal import SkyportalSender
+from mirar.processors.skyportal.skyportal_candidate import SkyportalCandidateUploader
+from mirar.processors.skyportal.skyportal_source import SkyportalSourceUploader
 from mirar.processors.skyportal.thumbnail import make_thumbnail
